Add GEOS heatflux and massflux updates.  This allows coupling to the GEOS coupled system where a semi-implicit thermodynamic coupling scheme is introduced. Similar to the explicit case, the fields fsurfn are provided by the coupler, along with their derivatives with respect to surface temperature dfsurfn_dTs. In this case, calc_Tsfc is still set to true, allowing ice surface and internal temperature to be updated implicitly. The resultant surface temperature change is passed back to the atmosphere model via coupler to complete the full update of its temperature profiles. This middle-ground approach, enabled by geos_heatflux=true, does not sacrifice accuracy because it does not need limiting effective conductivity as in the explicit case. In addition, in GEOS, the atmsophere model assumes vapor deposits or sublimates at 0 degC. In this case, mass conservation is enforced and the resulting discrepancy in energy is resolved by another term sblx and passed to ocean. This option is only on when geos_massflux=true.
- Add 4 new shortwave terms, uvrdr, uvrdf, pardr, pardf to the coupling.  These
terms represent a breakdown of the direct and diffuse visible shortwave terms
into two components, par = photosynthetical active radiation and uvr = rest of the
visible shortwave term.  The current visible shortwave is exactly represented by
these two components.  Includes adding atm forcing and terms associated with
radiation passthru to the ocean.

- Add calculation of GEOS heatflux.  In GEOS, surface and latent heat flux
is computed in the atmosphere at 0degC.  The sea ice model has to respect that
calculation, but then computes the d(dh)/dTs terms to correct the heatflux for
the sea ice temperature which is then applied conservatively in the coupled system.
Implementation includes turning off some of the heat flux calculations in Icepack.

- Add calculation of GEOS massflux term.  An equivalent correction is needed
to the mass and enthalpy terms to take into account the GEOS coupling.

- Add geos_heatflux and geos_massflux to parameters and namelist input

- Update discover port

- Update documentation